[CORE] Backport https://github.com/bitcoin/bitcoin/commit/d93b14dc5ddfb937b0cc18be425b9d048cefb66b
