+++ src/Makefile.am
@@ -81,6 +81,7 @@ BITCOIN_CORE_H = \
   accumulatorcheckpoints.h \
   accumulatorcheckpoints.json.h \
   accumulatormap.h \
+  addrdb.h \
   addrman.h \
   alert.h \
   allocators.h \
@@ -210,6 +211,7 @@ libbitcoin_util_a-clientversion.$(OBJEXT): obj/build.h
 libbitcoin_server_a_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES) $(MINIUPNPC_CPPFLAGS) $(EVENT_CFLAGS) $(EVENT_PTHREADS_CFLAGS)
 libbitcoin_server_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
 libbitcoin_server_a_SOURCES = \
+  addrdb.cpp \
   addrman.cpp \
   alert.cpp \
   bloom.cpp \
